Abstract database interface and sqlite implementation
