+++ src/colmap/controllers/incremental_pipeline_test.cc
@@ -30,6 +30,7 @@
 #include "colmap/controllers/incremental_pipeline.h"
 
 #include "colmap/estimators/alignment.h"
+#include "colmap/scene/database.h"
 #include "colmap/scene/synthetic.h"
 #include "colmap/util/testing.h"
 
@@ -71,7 +72,7 @@ void ExpectEqualReconstructions(const Reconstruction& gt,
 TEST(IncrementalPipeline, WithoutNoise) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.num_rigs = 2;
@@ -80,7 +81,8 @@ TEST(IncrementalPipeline, WithoutNoise) {
   synthetic_dataset_options.num_points3D = 50;
   synthetic_dataset_options.point2D_stddev = 0;
   synthetic_dataset_options.camera_has_prior_focal_length = false;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   auto reconstruction_manager = std::make_shared<ReconstructionManager>();
   IncrementalPipeline mapper(std::make_shared<IncrementalPipelineOptions>(),
@@ -100,7 +102,7 @@ TEST(IncrementalPipeline, WithoutNoise) {
 TEST(IncrementalPipeline, WithoutNoiseAndWithNonTrivialFrames) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.num_rigs = 2;
@@ -111,7 +113,8 @@ TEST(IncrementalPipeline, WithoutNoiseAndWithNonTrivialFrames) {
   synthetic_dataset_options.camera_has_prior_focal_length = false;
   synthetic_dataset_options.sensor_from_rig_translation_stddev = 0.05;
   synthetic_dataset_options.sensor_from_rig_rotation_stddev = 30;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   for (const bool refine_sensor_from_rig : {true, false}) {
     auto reconstruction_manager = std::make_shared<ReconstructionManager>();
@@ -135,7 +138,7 @@ TEST(IncrementalPipeline, WithoutNoiseAndWithNonTrivialFrames) {
 TEST(IncrementalPipeline, WithoutNoiseAndWithPanoramicNonTrivialFrames) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.num_rigs = 2;
@@ -146,7 +149,8 @@ TEST(IncrementalPipeline, WithoutNoiseAndWithPanoramicNonTrivialFrames) {
   synthetic_dataset_options.camera_has_prior_focal_length = false;
   synthetic_dataset_options.sensor_from_rig_translation_stddev = 0;
   synthetic_dataset_options.sensor_from_rig_rotation_stddev = 30;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   for (const bool refine_sensor_from_rig : {true, false}) {
     auto reconstruction_manager = std::make_shared<ReconstructionManager>();
@@ -170,7 +174,7 @@ TEST(IncrementalPipeline, WithoutNoiseAndWithPanoramicNonTrivialFrames) {
 TEST(IncrementalPipeline, WithPriorFocalLength) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.num_rigs = 2;
@@ -179,7 +183,8 @@ TEST(IncrementalPipeline, WithPriorFocalLength) {
   synthetic_dataset_options.num_points3D = 50;
   synthetic_dataset_options.point2D_stddev = 0;
   synthetic_dataset_options.camera_has_prior_focal_length = true;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   auto reconstruction_manager = std::make_shared<ReconstructionManager>();
   IncrementalPipeline mapper(std::make_shared<IncrementalPipelineOptions>(),
@@ -199,7 +204,7 @@ TEST(IncrementalPipeline, WithPriorFocalLength) {
 TEST(IncrementalPipeline, WithNoise) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.num_rigs = 2;
@@ -207,7 +212,8 @@ TEST(IncrementalPipeline, WithNoise) {
   synthetic_dataset_options.num_frames_per_rig = 7;
   synthetic_dataset_options.num_points3D = 100;
   synthetic_dataset_options.point2D_stddev = 0.5;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   auto reconstruction_manager = std::make_shared<ReconstructionManager>();
   IncrementalPipeline mapper(std::make_shared<IncrementalPipelineOptions>(),
@@ -227,7 +233,7 @@ TEST(IncrementalPipeline, WithNoise) {
 TEST(IncrementalPipeline, MultiReconstruction) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction1;
   Reconstruction gt_reconstruction2;
   SyntheticDatasetOptions synthetic_dataset_options;
@@ -236,9 +242,11 @@ TEST(IncrementalPipeline, MultiReconstruction) {
   synthetic_dataset_options.num_frames_per_rig = 5;
   synthetic_dataset_options.num_points3D = 50;
   synthetic_dataset_options.point2D_stddev = 0;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction1, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction1, database.get());
   synthetic_dataset_options.num_frames_per_rig = 4;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction2, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction2, database.get());
 
   auto reconstruction_manager = std::make_shared<ReconstructionManager>();
   auto mapper_options = std::make_shared<IncrementalPipelineOptions>();
@@ -274,7 +282,7 @@ TEST(IncrementalPipeline, MultiReconstruction) {
 TEST(IncrementalPipeline, FixExistingFrames) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.num_rigs = 1;
@@ -283,7 +291,8 @@ TEST(IncrementalPipeline, FixExistingFrames) {
   synthetic_dataset_options.num_points3D = 50;
   synthetic_dataset_options.point2D_stddev = 0;
   synthetic_dataset_options.camera_has_prior_focal_length = false;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   auto reconstruction_manager = std::make_shared<ReconstructionManager>();
   auto options = std::make_shared<IncrementalPipelineOptions>();
@@ -323,7 +332,7 @@ TEST(IncrementalPipeline, FixExistingFrames) {
 TEST(IncrementalPipeline, ChainedMatches) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.match_config =
@@ -333,7 +342,8 @@ TEST(IncrementalPipeline, ChainedMatches) {
   synthetic_dataset_options.num_frames_per_rig = 4;
   synthetic_dataset_options.num_points3D = 100;
   synthetic_dataset_options.point2D_stddev = 0;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   auto reconstruction_manager = std::make_shared<ReconstructionManager>();
   IncrementalPipeline mapper(std::make_shared<IncrementalPipelineOptions>(),
@@ -353,7 +363,7 @@ TEST(IncrementalPipeline, ChainedMatches) {
 TEST(IncrementalPipeline, PriorBasedSfMWithoutNoise) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.num_rigs = 2;
@@ -364,7 +374,8 @@ TEST(IncrementalPipeline, PriorBasedSfMWithoutNoise) {
 
   synthetic_dataset_options.use_prior_position = true;
   synthetic_dataset_options.prior_position_stddev = 0.0;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   std::shared_ptr<IncrementalPipelineOptions> mapper_options =
       std::make_shared<IncrementalPipelineOptions>();
@@ -392,7 +403,7 @@ TEST(IncrementalPipeline, PriorBasedSfMWithoutNoise) {
 TEST(IncrementalPipeline, PriorBasedSfMWithoutNoiseAndWithNonTrivialFrames) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.num_rigs = 2;
@@ -404,7 +415,8 @@ TEST(IncrementalPipeline, PriorBasedSfMWithoutNoiseAndWithNonTrivialFrames) {
 
   synthetic_dataset_options.use_prior_position = true;
   synthetic_dataset_options.prior_position_stddev = 0.0;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   std::shared_ptr<IncrementalPipelineOptions> mapper_options =
       std::make_shared<IncrementalPipelineOptions>();
@@ -431,7 +443,7 @@ TEST(IncrementalPipeline, PriorBasedSfMWithoutNoiseAndWithNonTrivialFrames) {
 TEST(IncrementalPipeline, PriorBasedSfMWithNoise) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.num_rigs = 2;
@@ -442,7 +454,8 @@ TEST(IncrementalPipeline, PriorBasedSfMWithNoise) {
 
   synthetic_dataset_options.use_prior_position = true;
   synthetic_dataset_options.prior_position_stddev = 1.5;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   std::shared_ptr<IncrementalPipelineOptions> mapper_options =
       std::make_shared<IncrementalPipelineOptions>();
@@ -469,7 +482,7 @@ TEST(IncrementalPipeline, PriorBasedSfMWithNoise) {
 TEST(IncrementalPipeline, GPSPriorBasedSfMWithNoise) {
   const std::string database_path = CreateTestDir() + "/database.db";
 
-  Database database(database_path);
+  auto database = Database::Open(database_path);
   Reconstruction gt_reconstruction;
   SyntheticDatasetOptions synthetic_dataset_options;
   synthetic_dataset_options.num_rigs = 2;
@@ -481,7 +494,8 @@ TEST(IncrementalPipeline, GPSPriorBasedSfMWithNoise) {
   synthetic_dataset_options.use_prior_position = true;
   synthetic_dataset_options.use_geographic_coords_prior = true;
   synthetic_dataset_options.prior_position_stddev = 1.5;
-  SynthesizeDataset(synthetic_dataset_options, &gt_reconstruction, &database);
+  SynthesizeDataset(
+      synthetic_dataset_options, &gt_reconstruction, database.get());
 
   std::shared_ptr<IncrementalPipelineOptions> mapper_options =
       std::make_shared<IncrementalPipelineOptions>();
