+++ src/colmap/exe/image.cc
@@ -279,11 +279,12 @@ int RunImageRegistrator(int argc, char** argv) {
     timer.Start();
     const size_t min_num_matches =
         static_cast<size_t>(options.mapper->min_num_matches);
-    database_cache = DatabaseCache::Create(Database(*options.database_path),
-                                           min_num_matches,
-                                           options.mapper->ignore_watermarks,
-                                           {options.mapper->image_names.begin(),
-                                            options.mapper->image_names.end()});
+    database_cache =
+        DatabaseCache::Create(*Database::Open(*options.database_path),
+                              min_num_matches,
+                              options.mapper->ignore_watermarks,
+                              {options.mapper->image_names.begin(),
+                               options.mapper->image_names.end()});
     timer.PrintMinutes();
   }
