+++ src/colmap/scene/database_cache_test.cc
@@ -29,61 +29,65 @@
 
 #include "colmap/scene/database_cache.h"
 
+#include "colmap/scene/database_sqlite.h"
+
 #include <gtest/gtest.h>
 
 namespace colmap {
 namespace {
 
-void CreateTestDatabase(Database& database) {
+std::shared_ptr<Database> CreateTestDatabase() {
+  auto database = Database::Open(kInMemorySqliteDatabasePath);
+
   Camera camera1 = Camera::CreateFromModelId(
       kInvalidCameraId, SimplePinholeCameraModel::model_id, 1, 1, 1);
-  camera1.camera_id = database.WriteCamera(camera1);
+  camera1.camera_id = database->WriteCamera(camera1);
   Camera camera2 = Camera::CreateFromModelId(
       kInvalidCameraId, SimplePinholeCameraModel::model_id, 2, 2, 2);
-  camera2.camera_id = database.WriteCamera(camera2);
+  camera2.camera_id = database->WriteCamera(camera2);
 
   Rig rig;
   rig.AddRefSensor(camera1.SensorId());
   rig.AddSensor(camera2.SensorId());
-  const rig_t rig_id = database.WriteRig(rig);
+  const rig_t rig_id = database->WriteRig(rig);
 
   Image image1;
   image1.SetName("image1");
   image1.SetCameraId(camera1.camera_id);
-  image1.SetImageId(database.WriteImage(image1));
+  image1.SetImageId(database->WriteImage(image1));
   Image image2;
   image2.SetName("image2");
   image2.SetCameraId(camera2.camera_id);
-  image2.SetImageId(database.WriteImage(image2));
+  image2.SetImageId(database->WriteImage(image2));
   Image image3;
   image3.SetName("image3");
   image3.SetCameraId(camera1.camera_id);
-  image3.SetImageId(database.WriteImage(image3));
+  image3.SetImageId(database->WriteImage(image3));
   Image image4;
   image4.SetName("image4");
   image4.SetCameraId(camera2.camera_id);
-  image4.SetImageId(database.WriteImage(image4));
+  image4.SetImageId(database->WriteImage(image4));
 
   Frame frame1;
   frame1.SetRigId(rig_id);
   frame1.AddDataId(image1.DataId());
   frame1.AddDataId(image2.DataId());
-  frame1.SetFrameId(database.WriteFrame(frame1));
+  frame1.SetFrameId(database->WriteFrame(frame1));
   Frame frame2;
   frame2.SetRigId(rig_id);
   frame2.AddDataId(image3.DataId());
   frame2.AddDataId(image4.DataId());
-  frame2.SetFrameId(database.WriteFrame(frame2));
+  frame2.SetFrameId(database->WriteFrame(frame2));
 
-  database.WritePosePrior(image1.ImageId(),
-                          PosePrior(Eigen::Vector3d::Random()));
-  database.WritePosePrior(image2.ImageId(),
-                          PosePrior(Eigen::Vector3d::Random()));
+  database->WritePosePrior(image1.ImageId(),
+                           PosePrior(Eigen::Vector3d::Random()));
+  database->WritePosePrior(image2.ImageId(),
+                           PosePrior(Eigen::Vector3d::Random()));
 
-  database.WriteKeypoints(image1.ImageId(), FeatureKeypoints(10));
-  database.WriteKeypoints(image2.ImageId(), FeatureKeypoints(5));
-  database.WriteKeypoints(image3.ImageId(), FeatureKeypoints(6));
-  database.WriteKeypoints(image4.ImageId(), FeatureKeypoints(3));
+  database->WriteKeypoints(image1.ImageId(), FeatureKeypoints(10));
+  database->WriteKeypoints(image2.ImageId(), FeatureKeypoints(5));
+  database->WriteKeypoints(image3.ImageId(), FeatureKeypoints(6));
+  database->WriteKeypoints(image4.ImageId(), FeatureKeypoints(3));
 
   TwoViewGeometry two_view_geometry;
   two_view_geometry.inlier_matches = {{0, 1}};
@@ -94,12 +98,14 @@ void CreateTestDatabase(Database& database) {
   two_view_geometry.H = Eigen::Matrix3d::Random();
   two_view_geometry.cam2_from_cam1 =
       Rigid3d(Eigen::Quaterniond::UnitRandom(), Eigen::Vector3d::Random());
-  database.WriteTwoViewGeometry(
+  database->WriteTwoViewGeometry(
       image1.ImageId(), image2.ImageId(), two_view_geometry);
-  database.WriteTwoViewGeometry(
+  database->WriteTwoViewGeometry(
       image2.ImageId(), image3.ImageId(), two_view_geometry);
-  database.WriteTwoViewGeometry(
+  database->WriteTwoViewGeometry(
       image3.ImageId(), image4.ImageId(), two_view_geometry);
+
+  return database;
 }
 
 TEST(DatabaseCache, Empty) {
@@ -112,9 +118,8 @@ TEST(DatabaseCache, Empty) {
 }
 
 TEST(DatabaseCache, ConstructFromDatabase) {
-  Database database(Database::kInMemoryDatabasePath);
-  CreateTestDatabase(database);
-  auto cache = DatabaseCache::Create(database,
+  auto database = CreateTestDatabase();
+  auto cache = DatabaseCache::Create(*database,
                                      /*min_num_matches=*/0,
                                      /*ignore_watermarks=*/false,
                                      /*image_names=*/{});
@@ -125,22 +130,22 @@ TEST(DatabaseCache, ConstructFromDatabase) {
   EXPECT_EQ(cache->NumImages(), 4);
   EXPECT_EQ(cache->NumPosePriors(), 2);
 
-  for (const Rig& rig : database.ReadAllRigs()) {
+  for (const Rig& rig : database->ReadAllRigs()) {
     EXPECT_TRUE(cache->ExistsRig(rig.RigId()));
     EXPECT_EQ(cache->Rig(rig.RigId()), rig);
   }
 
-  for (const Camera& camera : database.ReadAllCameras()) {
+  for (const Camera& camera : database->ReadAllCameras()) {
     EXPECT_TRUE(cache->ExistsCamera(camera.camera_id));
     EXPECT_EQ(cache->Camera(camera.camera_id), camera);
   }
 
-  for (const Frame& frame : database.ReadAllFrames()) {
+  for (const Frame& frame : database->ReadAllFrames()) {
     EXPECT_TRUE(cache->ExistsFrame(frame.FrameId()));
     EXPECT_EQ(cache->Frame(frame.FrameId()), frame);
   }
 
-  const std::vector<Image> images = database.ReadAllImages();
+  const std::vector<Image> images = database->ReadAllImages();
   EXPECT_TRUE(cache->ExistsImage(images[0].ImageId()));
   EXPECT_EQ(cache->Image(images[0].ImageId()).NumPoints2D(), 10);
   EXPECT_TRUE(cache->ExistsImage(images[1].ImageId()));
@@ -179,13 +184,12 @@ TEST(DatabaseCache, ConstructFromDatabase) {
 }
 
 TEST(DatabaseCache, ConstructFromDatabaseWithCustomImages) {
-  Database database(Database::kInMemoryDatabasePath);
-  CreateTestDatabase(database);
+  auto database = CreateTestDatabase();
 
   // Note that the first two images are part of the same frame.
-  const std::vector<Image> images = database.ReadAllImages();
+  const std::vector<Image> images = database->ReadAllImages();
   auto cache = DatabaseCache::Create(
-      database,
+      *database,
       /*min_num_matches=*/0,
       /*ignore_watermarks=*/false,
       /*image_names=*/{images[0].Name(), images[1].Name()});
@@ -209,10 +213,12 @@ TEST(DatabaseCache, ConstructFromDatabaseWithCustomImages) {
             1);
 }
 
-void CreateLegacyTestDatabase(Database& database) {
+std::shared_ptr<Database> CreateLegacyTestDatabase() {
+  auto database = Database::Open(kInMemorySqliteDatabasePath);
+
   const Camera camera = Camera::CreateFromModelId(
       kInvalidCameraId, SimplePinholeCameraModel::model_id, 1, 1, 1);
-  const camera_t camera_id = database.WriteCamera(camera);
+  const camera_t camera_id = database->WriteCamera(camera);
   Image image1;
   image1.SetName("image1");
   image1.SetCameraId(camera_id);
@@ -222,15 +228,15 @@ void CreateLegacyTestDatabase(Database& database) {
   Image image3;
   image3.SetName("image3");
   image3.SetCameraId(camera_id);
-  const image_t image_id1 = database.WriteImage(image1);
-  const image_t image_id2 = database.WriteImage(image2);
-  const image_t image_id3 = database.WriteImage(image3);
-  database.WritePosePrior(image_id1, PosePrior(Eigen::Vector3d::Random()));
-  database.WritePosePrior(image_id2, PosePrior(Eigen::Vector3d::Random()));
-  database.WritePosePrior(image_id3, PosePrior(Eigen::Vector3d::Random()));
-  database.WriteKeypoints(image_id1, FeatureKeypoints(10));
-  database.WriteKeypoints(image_id2, FeatureKeypoints(5));
-  database.WriteKeypoints(image_id3, FeatureKeypoints(7));
+  const image_t image_id1 = database->WriteImage(image1);
+  const image_t image_id2 = database->WriteImage(image2);
+  const image_t image_id3 = database->WriteImage(image3);
+  database->WritePosePrior(image_id1, PosePrior(Eigen::Vector3d::Random()));
+  database->WritePosePrior(image_id2, PosePrior(Eigen::Vector3d::Random()));
+  database->WritePosePrior(image_id3, PosePrior(Eigen::Vector3d::Random()));
+  database->WriteKeypoints(image_id1, FeatureKeypoints(10));
+  database->WriteKeypoints(image_id2, FeatureKeypoints(5));
+  database->WriteKeypoints(image_id3, FeatureKeypoints(7));
   TwoViewGeometry two_view_geometry;
   two_view_geometry.inlier_matches = {{0, 1}};
   two_view_geometry.config =
@@ -240,14 +246,15 @@ void CreateLegacyTestDatabase(Database& database) {
   two_view_geometry.H = Eigen::Matrix3d::Random();
   two_view_geometry.cam2_from_cam1 =
       Rigid3d(Eigen::Quaterniond::UnitRandom(), Eigen::Vector3d::Random());
-  database.WriteTwoViewGeometry(image_id1, image_id2, two_view_geometry);
-  database.WriteTwoViewGeometry(image_id1, image_id3, two_view_geometry);
+  database->WriteTwoViewGeometry(image_id1, image_id2, two_view_geometry);
+  database->WriteTwoViewGeometry(image_id1, image_id3, two_view_geometry);
+
+  return database;
 }
 
 TEST(DatabaseCache, ConstructFromLegacyDatabaseWithoutRigsAndFrames) {
-  Database database(Database::kInMemoryDatabasePath);
-  CreateLegacyTestDatabase(database);
-  auto cache = DatabaseCache::Create(database,
+  auto database = CreateLegacyTestDatabase();
+  auto cache = DatabaseCache::Create(*database,
                                      /*min_num_matches=*/0,
                                      /*ignore_watermarks=*/false,
                                      /*image_names=*/{});
@@ -278,11 +285,10 @@ TEST(DatabaseCache, ConstructFromLegacyDatabaseWithoutRigsAndFrames) {
 }
 
 TEST(DatabaseCache, ConstructFromLegacyDatabaseWithCustomImages) {
-  Database database(Database::kInMemoryDatabasePath);
-  CreateLegacyTestDatabase(database);
-  const std::vector<Image> images = database.ReadAllImages();
+  auto database = CreateLegacyTestDatabase();
+  const std::vector<Image> images = database->ReadAllImages();
   auto cache = DatabaseCache::Create(
-      database,
+      *database,
       /*min_num_matches=*/0,
       /*ignore_watermarks=*/false,
       /*image_names=*/{images[0].Name(), images[2].Name()});
