+++ src/colmap/scene/database_sqlite.h
@@ -29,43 +29,16 @@
 
 #pragma once
 
-#include "colmap/util/logging.h"
+#include "colmap/scene/database.h"
 
-#include <cstdio>
-#include <cstdlib>
+#include <memory>
 #include <string>
-
-#include <sqlite3.h>
 
 namespace colmap {
 
-inline int SQLite3CallHelper(int result_code,
-                             const std::string& filename,
-                             int line) {
-  switch (result_code) {
-    case SQLITE_OK:
-    case SQLITE_ROW:
-    case SQLITE_DONE:
-      return result_code;
-    default:
-      LogMessageFatalThrow<std::runtime_error>(filename.c_str(), line).stream()
-          << "SQLite error: " << sqlite3_errstr(result_code);
-      return result_code;
-  }
-}
+// Can be used to construct temporary in-memory database.
+constexpr inline char kInMemorySqliteDatabasePath[] = ":memory:";
 
-#define SQLITE3_CALL(func) SQLite3CallHelper(func, __FILE__, __LINE__)
-
-#define SQLITE3_EXEC(database, sql, callback)                             \
-  {                                                                       \
-    char* err_msg = nullptr;                                              \
-    const int result_code = sqlite3_exec(                                 \
-        THROW_CHECK_NOTNULL(database), sql, callback, nullptr, &err_msg); \
-    if (result_code != SQLITE_OK) {                                       \
-      LOG(ERROR) << "SQLite error [" << __FILE__ << ", line " << __LINE__ \
-                 << "]: " << err_msg;                                     \
-      sqlite3_free(err_msg);                                              \
-    }                                                                     \
-  }
+std::shared_ptr<Database> OpenSqliteDatabase(const std::string& path);
 
 }  // namespace colmap
