+++ src/colmap/scene/synthetic_test.cc
@@ -30,6 +30,7 @@
 #include "colmap/scene/synthetic.h"
 
 #include "colmap/geometry/triangulation.h"
+#include "colmap/scene/database_sqlite.h"
 #include "colmap/scene/projection.h"
 #include "colmap/util/file.h"
 #include "colmap/util/testing.h"
@@ -40,35 +41,35 @@ namespace colmap {
 namespace {
 
 TEST(SynthesizeDataset, Nominal) {
-  Database database(Database::kInMemoryDatabasePath);
+  auto database = Database::Open(kInMemorySqliteDatabasePath);
   Reconstruction reconstruction;
   SyntheticDatasetOptions options;
   options.num_rigs = 2;
   options.num_cameras_per_rig = 3;
   options.num_frames_per_rig = 3;
-  SynthesizeDataset(options, &reconstruction, &database);
+  SynthesizeDataset(options, &reconstruction, database.get());
 
   const std::string test_dir = CreateTestDir();
   const std::string sparse_path = test_dir + "/sparse";
   CreateDirIfNotExists(sparse_path);
   reconstruction.Write(sparse_path);
 
-  EXPECT_EQ(database.NumRigs(), options.num_rigs);
+  EXPECT_EQ(database->NumRigs(), options.num_rigs);
   EXPECT_EQ(reconstruction.NumRigs(), options.num_rigs);
   for (const auto& rig : reconstruction.Rigs()) {
     EXPECT_GE(rig.second.NumSensors(), options.num_cameras_per_rig);
   }
 
-  EXPECT_EQ(database.NumCameras(),
+  EXPECT_EQ(database->NumCameras(),
             options.num_rigs * options.num_cameras_per_rig);
   EXPECT_EQ(reconstruction.NumCameras(),
             options.num_rigs * options.num_cameras_per_rig);
   for (const auto& [camera_id, camera] : reconstruction.Cameras()) {
-    EXPECT_EQ(camera, database.ReadCamera(camera_id));
+    EXPECT_EQ(camera, database->ReadCamera(camera_id));
     EXPECT_EQ(camera.model_id, options.camera_model_id);
   }
 
-  EXPECT_EQ(database.NumFrames(),
+  EXPECT_EQ(database->NumFrames(),
             options.num_rigs * options.num_frames_per_rig);
   EXPECT_EQ(reconstruction.NumFrames(),
             options.num_rigs * options.num_frames_per_rig);
@@ -76,12 +77,12 @@ TEST(SynthesizeDataset, Nominal) {
     Frame reconstruction_frame = frame;
     EXPECT_TRUE(reconstruction_frame.HasPose());
     reconstruction_frame.ResetPose();
-    EXPECT_EQ(reconstruction_frame, database.ReadFrame(frame_id));
+    EXPECT_EQ(reconstruction_frame, database->ReadFrame(frame_id));
     EXPECT_EQ(reconstruction_frame.DataIds().size(),
               reconstruction_frame.RigPtr()->NumSensors());
   }
 
-  EXPECT_EQ(database.NumImages(),
+  EXPECT_EQ(database->NumImages(),
             options.num_rigs * options.num_cameras_per_rig *
                 options.num_frames_per_rig);
   EXPECT_EQ(reconstruction.NumImages(),
@@ -91,13 +92,13 @@ TEST(SynthesizeDataset, Nominal) {
             options.num_rigs * options.num_frames_per_rig);
   std::set<std::string> image_names;
   for (const auto& image : reconstruction.Images()) {
-    EXPECT_EQ(image.second.Name(), database.ReadImage(image.first).Name());
+    EXPECT_EQ(image.second.Name(), database->ReadImage(image.first).Name());
     image_names.insert(image.second.Name());
     EXPECT_EQ(image.second.NumPoints2D(),
-              database.ReadKeypoints(image.first).size());
+              database->ReadKeypoints(image.first).size());
     EXPECT_EQ(image.second.NumPoints2D(),
-              database.ReadDescriptors(image.first).rows());
-    EXPECT_EQ(database.ReadDescriptors(image.first).cols(), 128);
+              database->ReadDescriptors(image.first).rows());
+    EXPECT_EQ(database->ReadDescriptors(image.first).cols(), 128);
     EXPECT_EQ(image.second.NumPoints2D(),
               options.num_points3D + options.num_points2D_without_point3D);
     EXPECT_EQ(image.second.NumPoints3D(), options.num_points3D);
@@ -106,8 +107,8 @@ TEST(SynthesizeDataset, Nominal) {
 
   const int num_image_pairs =
       reconstruction.NumImages() * (reconstruction.NumImages() - 1) / 2;
-  EXPECT_EQ(database.NumVerifiedImagePairs(), num_image_pairs);
-  EXPECT_EQ(database.NumInlierMatches(),
+  EXPECT_EQ(database->NumVerifiedImagePairs(), num_image_pairs);
+  EXPECT_EQ(database->NumInlierMatches(),
             num_image_pairs * options.num_points3D);
 
   EXPECT_NEAR(reconstruction.ComputeMeanReprojectionError(), 0, 1e-6);
@@ -129,7 +130,7 @@ TEST(SynthesizeDataset, Nominal) {
 
     // Make sure all descriptors of the same 3D point have identical features.
     const FeatureDescriptor descriptors =
-        database.ReadDescriptors(point3D.track.Element(0).image_id)
+        database->ReadDescriptors(point3D.track.Element(0).image_id)
             .row(point3D.track.Element(0).point2D_idx);
 
     double max_tri_angle = 0;
@@ -143,7 +144,7 @@ TEST(SynthesizeDataset, Nominal) {
           point2D.xy, point3D.xyz, image1.CamFromWorld(), camera1);
       EXPECT_LE(squared_reproj_error, kMaxReprojError * kMaxReprojError);
       EXPECT_EQ(descriptors,
-                database.ReadDescriptors(point3D.track.Element(i1).image_id)
+                database->ReadDescriptors(point3D.track.Element(i1).image_id)
                     .row(point3D.track.Element(i1).point2D_idx));
 
       Eigen::Vector3d proj_center1;
@@ -168,44 +169,44 @@ TEST(SynthesizeDataset, Nominal) {
 }
 
 TEST(SynthesizeDataset, MultipleTimes) {
-  Database database(Database::kInMemoryDatabasePath);
+  auto database = Database::Open(kInMemorySqliteDatabasePath);
   Reconstruction reconstruction;
   SyntheticDatasetOptions options;
   options.num_rigs = 2;
   options.num_cameras_per_rig = 3;
   options.num_frames_per_rig = 3;
-  SynthesizeDataset(options, &reconstruction, &database);
-  SynthesizeDataset(options, &reconstruction, &database);
+  SynthesizeDataset(options, &reconstruction, database.get());
+  SynthesizeDataset(options, &reconstruction, database.get());
 
-  EXPECT_EQ(database.NumRigs(), 2 * options.num_rigs);
-  EXPECT_EQ(reconstruction.NumRigs(), database.NumRigs());
+  EXPECT_EQ(database->NumRigs(), 2 * options.num_rigs);
+  EXPECT_EQ(reconstruction.NumRigs(), database->NumRigs());
 
-  EXPECT_EQ(database.NumCameras(),
+  EXPECT_EQ(database->NumCameras(),
             2 * options.num_rigs * options.num_cameras_per_rig);
-  EXPECT_EQ(reconstruction.NumCameras(), database.NumCameras());
+  EXPECT_EQ(reconstruction.NumCameras(), database->NumCameras());
 
-  EXPECT_EQ(database.NumFrames(),
+  EXPECT_EQ(database->NumFrames(),
             2 * options.num_rigs * options.num_frames_per_rig);
-  EXPECT_EQ(database.NumFrames(), reconstruction.NumFrames());
+  EXPECT_EQ(database->NumFrames(), reconstruction.NumFrames());
 
-  EXPECT_EQ(database.NumImages(),
+  EXPECT_EQ(database->NumImages(),
             2 * options.num_rigs * options.num_cameras_per_rig *
                 options.num_frames_per_rig);
-  EXPECT_EQ(database.NumImages(), reconstruction.NumImages());
+  EXPECT_EQ(database->NumImages(), reconstruction.NumImages());
 
   const int num_image_pairs =
       reconstruction.NumImages() * (reconstruction.NumImages() - 1) / 2;
-  EXPECT_EQ(database.NumVerifiedImagePairs(), num_image_pairs);
+  EXPECT_EQ(database->NumVerifiedImagePairs(), num_image_pairs);
 
   EXPECT_EQ(reconstruction.NumPoints3D(), 2 * options.num_points3D);
 }
 
 TEST(SynthesizeDataset, WithNoise) {
-  Database database(Database::kInMemoryDatabasePath);
+  auto database = Database::Open(kInMemorySqliteDatabasePath);
   Reconstruction reconstruction;
   SyntheticDatasetOptions options;
   options.point2D_stddev = 2.0;
-  SynthesizeDataset(options, &reconstruction, &database);
+  SynthesizeDataset(options, &reconstruction, database.get());
 
   EXPECT_NEAR(reconstruction.ComputeMeanReprojectionError(),
               options.point2D_stddev,
@@ -215,17 +216,17 @@ TEST(SynthesizeDataset, WithNoise) {
 }
 
 TEST(SynthesizeDataset, WithPriors) {
-  Database database(Database::kInMemoryDatabasePath);
+  auto database = Database::Open(kInMemorySqliteDatabasePath);
   Reconstruction reconstruction;
   SyntheticDatasetOptions options;
   options.use_prior_position = true;
   options.prior_position_stddev = 0.;
-  SynthesizeDataset(options, &reconstruction, &database);
+  SynthesizeDataset(options, &reconstruction, database.get());
 
   for (const auto& image : reconstruction.Images()) {
-    if (database.ExistsPosePrior(image.first)) {
+    if (database->ExistsPosePrior(image.first)) {
       EXPECT_NEAR((image.second.ProjectionCenter() -
-                   database.ReadPosePrior(image.first).position)
+                   database->ReadPosePrior(image.first).position)
                       .norm(),
                   0.,
                   1e-9);
@@ -234,71 +235,71 @@ TEST(SynthesizeDataset, WithPriors) {
 }
 
 TEST(SynthesizeDataset, MultiReconstruction) {
-  Database database(Database::kInMemoryDatabasePath);
+  auto database = Database::Open(kInMemorySqliteDatabasePath);
   Reconstruction reconstruction1;
   Reconstruction reconstruction2;
   SyntheticDatasetOptions options;
-  SynthesizeDataset(options, &reconstruction1, &database);
-  SynthesizeDataset(options, &reconstruction2, &database);
+  SynthesizeDataset(options, &reconstruction1, database.get());
+  SynthesizeDataset(options, &reconstruction2, database.get());
 
   const int num_cameras = options.num_rigs * options.num_cameras_per_rig;
-  EXPECT_EQ(database.NumCameras(), 2 * num_cameras);
+  EXPECT_EQ(database->NumCameras(), 2 * num_cameras);
   EXPECT_EQ(reconstruction1.NumCameras(), num_cameras);
   EXPECT_EQ(reconstruction1.NumCameras(), num_cameras);
   const int num_images = num_cameras * options.num_frames_per_rig;
-  EXPECT_EQ(database.NumImages(), 2 * num_images);
+  EXPECT_EQ(database->NumImages(), 2 * num_images);
   EXPECT_EQ(reconstruction1.NumImages(), num_images);
   EXPECT_EQ(reconstruction2.NumImages(), num_images);
   EXPECT_EQ(reconstruction1.NumRegFrames(), num_images);
   EXPECT_EQ(reconstruction2.NumRegFrames(), num_images);
   const int num_image_pairs = num_images * (num_images - 1) / 2;
-  EXPECT_EQ(database.NumVerifiedImagePairs(), 2 * num_image_pairs);
-  EXPECT_EQ(database.NumInlierMatches(),
+  EXPECT_EQ(database->NumVerifiedImagePairs(), 2 * num_image_pairs);
+  EXPECT_EQ(database->NumInlierMatches(),
             2 * num_image_pairs * options.num_points3D);
 }
 
 TEST(SynthesizeDataset, ExhaustiveMatches) {
-  Database database(Database::kInMemoryDatabasePath);
+  auto database = Database::Open(kInMemorySqliteDatabasePath);
   Reconstruction reconstruction;
   SyntheticDatasetOptions options;
   options.match_config = SyntheticDatasetOptions::MatchConfig::EXHAUSTIVE;
-  SynthesizeDataset(options, &reconstruction, &database);
+  SynthesizeDataset(options, &reconstruction, database.get());
 
   const int num_images = options.num_rigs * options.num_cameras_per_rig *
                          options.num_frames_per_rig;
   const int num_image_pairs = num_images * (num_images - 1) / 2;
-  EXPECT_EQ(database.NumMatchedImagePairs(), num_image_pairs);
-  EXPECT_EQ(database.NumVerifiedImagePairs(), num_image_pairs);
-  EXPECT_EQ(database.NumInlierMatches(),
+  EXPECT_EQ(database->NumMatchedImagePairs(), num_image_pairs);
+  EXPECT_EQ(database->NumVerifiedImagePairs(), num_image_pairs);
+  EXPECT_EQ(database->NumInlierMatches(),
             num_image_pairs * options.num_points3D);
 }
 
 TEST(SynthesizeDataset, ChainedMatches) {
-  Database database(Database::kInMemoryDatabasePath);
+  auto database = Database::Open(kInMemorySqliteDatabasePath);
   Reconstruction reconstruction;
   SyntheticDatasetOptions options;
   options.match_config = SyntheticDatasetOptions::MatchConfig::CHAINED;
-  SynthesizeDataset(options, &reconstruction, &database);
+  SynthesizeDataset(options, &reconstruction, database.get());
 
   const int num_images = options.num_rigs * options.num_cameras_per_rig *
                          options.num_frames_per_rig;
   const int num_image_pairs = num_images - 1;
-  EXPECT_EQ(database.NumMatchedImagePairs(), num_image_pairs);
-  EXPECT_EQ(database.NumVerifiedImagePairs(), num_image_pairs);
-  EXPECT_EQ(database.NumInlierMatches(),
+  EXPECT_EQ(database->NumMatchedImagePairs(), num_image_pairs);
+  EXPECT_EQ(database->NumVerifiedImagePairs(), num_image_pairs);
+  EXPECT_EQ(database->NumInlierMatches(),
             num_image_pairs * options.num_points3D);
-  for (const auto& [pair_id, _] : database.ReadAllMatches()) {
+  for (const auto& [pair_id, _] : database->ReadAllMatches()) {
     const auto [image_id1, image_id2] = Database::PairIdToImagePair(pair_id);
     EXPECT_EQ(image_id1 + 1, image_id2);
   }
-  for (const auto& [pair_id, _] : database.ReadTwoViewGeometries()) {
+  for (const auto& [pair_id, _] : database->ReadTwoViewGeometries()) {
     const auto [image_id1, image_id2] = Database::PairIdToImagePair(pair_id);
     EXPECT_EQ(image_id1 + 1, image_id2);
   }
 }
 
 TEST(SynthesizeDataset, NoDatabase) {
-  Database database(Database::kInMemoryDatabasePath);
+  auto database = Database::Open(kInMemorySqliteDatabasePath);
   SyntheticDatasetOptions options;
   Reconstruction reconstruction;
   SynthesizeDataset(options, &reconstruction);
