+++ src/colmap/ui/database_management_widget.cc
@@ -29,17 +29,16 @@
 
 #include "colmap/ui/database_management_widget.h"
 
+#include "colmap/scene/database.h"
 #include "colmap/sensor/models.h"
 #include "colmap/util/file.h"
+#include "colmap/util/misc.h"
 
 namespace colmap {
 
-TwoViewInfoTab::TwoViewInfoTab(QWidget* parent,
-                               OptionManager* options,
-                               Database* database)
+TwoViewInfoTab::TwoViewInfoTab(QWidget* parent, OptionManager* options)
     : QWidget(parent),
       options_(options),
-      database_(database),
       matches_viewer_widget_(new FeatureImageViewerWidget(parent, "matches")) {}
 
 void TwoViewInfoTab::Clear() {
@@ -145,30 +144,29 @@ void TwoViewInfoTab::FillTable() {
   table_widget_->resizeColumnsToContents();
 }
 
-MatchesTab::MatchesTab(QWidget* parent,
-                       OptionManager* options,
-                       Database* database)
-    : TwoViewInfoTab(parent, options, database) {
+MatchesTab::MatchesTab(QWidget* parent, OptionManager* options)
+    : TwoViewInfoTab(parent, options) {
   QStringList table_header;
   table_header << "image_id"
                << "num_matches";
   InitializeTable(table_header);
 }
 
-void MatchesTab::Reload(const std::vector<Image>& images,
+void MatchesTab::Reload(const std::shared_ptr<Database>& database,
+                        const std::vector<Image>& images,
                         const image_t image_id) {
+  database_ = database;
   matches_.clear();
 
   // Find all matched images
-
   for (const auto& image : images) {
     if (image.ImageId() == image_id) {
       image_ = &image;
       continue;
     }
 
-    if (database_->ExistsMatches(image_id, image.ImageId())) {
-      const auto matches = database_->ReadMatches(image_id, image.ImageId());
+    if (database->ExistsMatches(image_id, image.ImageId())) {
+      const auto matches = database->ReadMatches(image_id, image.ImageId());
 
       if (matches.size() > 0) {
         matches_.emplace_back(&image, matches);
@@ -180,9 +178,8 @@ void MatchesTab::Reload(const std::vector<Image>& images,
 }
 
 TwoViewGeometriesTab::TwoViewGeometriesTab(QWidget* parent,
-                                           OptionManager* options,
-                                           Database* database)
-    : TwoViewInfoTab(parent, options, database) {
+                                           OptionManager* options)
+    : TwoViewInfoTab(parent, options) {
   QStringList table_header;
   table_header << "image_id"
                << "num_matches"
@@ -190,8 +187,10 @@ TwoViewGeometriesTab::TwoViewGeometriesTab(QWidget* parent,
   InitializeTable(table_header);
 }
 
-void TwoViewGeometriesTab::Reload(const std::vector<Image>& images,
+void TwoViewGeometriesTab::Reload(const std::shared_ptr<Database>& database,
+                                  const std::vector<Image>& images,
                                   const image_t image_id) {
+  database_ = database;
   matches_.clear();
   configs_.clear();
 
@@ -203,11 +202,10 @@ void TwoViewGeometriesTab::Reload(const std::vector<Image>& images,
       continue;
     }
 
-    if (database_->ExistsInlierMatches(image_id, image.ImageId())) {
+    if (database->ExistsInlierMatches(image_id, image.ImageId())) {
       const auto two_view_geometry =
-          database_->ReadTwoViewGeometry(image_id, image.ImageId());
-
-      if (two_view_geometry.inlier_matches.size() > 0) {
+          database->ReadTwoViewGeometry(image_id, image.ImageId());
+      if (!two_view_geometry.inlier_matches.empty()) {
         matches_.emplace_back(&image, two_view_geometry.inlier_matches);
         configs_.push_back(two_view_geometry.config);
       }
@@ -218,9 +216,8 @@ void TwoViewGeometriesTab::Reload(const std::vector<Image>& images,
 }
 
 OverlappingImagesWidget::OverlappingImagesWidget(QWidget* parent,
-                                                 OptionManager* options,
-                                                 Database* database)
-    : parent_(parent), options_(options) {
+                                                 OptionManager* options)
+    : parent_(parent) {
   // Do not change flag, to make sure feature database is not accessed from
   // multiple threads.
   setWindowFlags(Qt::Window);
@@ -230,10 +227,10 @@ OverlappingImagesWidget::OverlappingImagesWidget(QWidget* parent,
 
   tab_widget_ = new QTabWidget(this);
 
-  matches_tab_ = new MatchesTab(this, options_, database);
+  matches_tab_ = new MatchesTab(this, options);
   tab_widget_->addTab(matches_tab_, tr("Matches"));
 
-  two_view_geometries_tab_ = new TwoViewGeometriesTab(this, options_, database);
+  two_view_geometries_tab_ = new TwoViewGeometriesTab(this, options);
   tab_widget_->addTab(two_view_geometries_tab_, tr("Two-view geometries"));
 
   grid->addWidget(tab_widget_, 0, 0);
@@ -246,15 +243,17 @@ OverlappingImagesWidget::OverlappingImagesWidget(QWidget* parent,
   grid->addWidget(close_button, 1, 0, Qt::AlignRight);
 }
 
-void OverlappingImagesWidget::ShowMatches(const std::vector<Image>& images,
-                                          const image_t image_id) {
+void OverlappingImagesWidget::ShowMatches(
+    const std::shared_ptr<Database>& database,
+    const std::vector<Image>& images,
+    const image_t image_id) {
   parent_->setDisabled(true);
 
   setWindowTitle(
       QString::fromStdString("Matches for image " + std::to_string(image_id)));
 
-  matches_tab_->Reload(images, image_id);
-  two_view_geometries_tab_->Reload(images, image_id);
+  matches_tab_->Reload(database, images, image_id);
+  two_view_geometries_tab_->Reload(database, images, image_id);
 }
 
 void OverlappingImagesWidget::closeEvent(QCloseEvent*) {
@@ -263,8 +262,7 @@ void OverlappingImagesWidget::closeEvent(QCloseEvent*) {
   parent_->setEnabled(true);
 }
 
-CameraTab::CameraTab(QWidget* parent, Database* database)
-    : QWidget(parent), database_(database) {
+CameraTab::CameraTab(QWidget* parent) : QWidget(parent) {
   QGridLayout* grid = new QGridLayout(this);
 
   info_label_ = new QLabel(this);
@@ -304,12 +302,14 @@ CameraTab::CameraTab(QWidget* parent, Database* database)
   grid->setColumnStretch(0, 1);
 }
 
-void CameraTab::Reload() {
+void CameraTab::Reload(const std::shared_ptr<Database>& database) {
+  database_ = database;
+
   QString info;
-  info += QString("Cameras: ") + QString::number(database_->NumCameras());
+  info += QString("Cameras: ") + QString::number(database->NumCameras());
   info_label_->setText(info);
 
-  cameras_ = database_->ReadAllCameras();
+  cameras_ = database->ReadAllCameras();
 
   // Make sure, itemChanged is not invoked, while setting up the table.
   table_widget_->blockSignals(true);
@@ -418,7 +418,7 @@ void CameraTab::Add() {
   database_->WriteCamera(camera);
 
   // Reload all cameras
-  Reload();
+  Reload(database_);
 
   // Highlight new camera
   table_widget_->selectRow(cameras_.size() - 1);
@@ -461,17 +461,13 @@ void CameraTab::SetModel() {
 
   table_widget_->blockSignals(false);
 
-  Reload();
+  Reload(database_);
 }
 
 ImageTab::ImageTab(QWidget* parent,
                    CameraTab* camera_tab,
-                   OptionManager* options,
-                   Database* database)
-    : QWidget(parent),
-      camera_tab_(camera_tab),
-      options_(options),
-      database_(database) {
+                   OptionManager* options)
+    : QWidget(parent), camera_tab_(camera_tab), options_(options) {
   QGridLayout* grid = new QGridLayout(this);
 
   info_label_ = new QLabel(this);
@@ -525,18 +521,19 @@ ImageTab::ImageTab(QWidget* parent,
   grid->setColumnStretch(0, 3);
 
   image_viewer_widget_ = new FeatureImageViewerWidget(parent, "keypoints");
-  overlapping_images_widget_ =
-      new OverlappingImagesWidget(parent, options, database_);
+  overlapping_images_widget_ = new OverlappingImagesWidget(parent, options);
 }
 
-void ImageTab::Reload() {
+void ImageTab::Reload(const std::shared_ptr<Database>& database) {
+  database_ = database;
+
   QString info;
-  info += QString("Images: ") + QString::number(database_->NumImages());
+  info += QString("Images: ") + QString::number(database->NumImages());
   info += QString('\n');
-  info += QString("Features: ") + QString::number(database_->NumKeypoints());
+  info += QString("Features: ") + QString::number(database->NumKeypoints());
   info_label_->setText(info);
 
-  images_ = database_->ReadAllImages();
+  images_ = database->ReadAllImages();
 
   // Make sure, itemChanged is not invoked, while setting up the table
   table_widget_->blockSignals(true);
@@ -631,7 +628,7 @@ void ImageTab::ShowMatches() {
 
   const auto& image = images_[select->selectedRows().begin()->row()];
 
-  overlapping_images_widget_->ShowMatches(images_, image.ImageId());
+  overlapping_images_widget_->ShowMatches(database_, images_, image.ImageId());
   overlapping_images_widget_->show();
   overlapping_images_widget_->raise();
 }
@@ -707,11 +704,10 @@ void ImageTab::SplitCamera() {
 
   table_widget_->blockSignals(false);
 
-  camera_tab_->Reload();
+  camera_tab_->Reload(database_);
 }
 
-PosePriorsTab::PosePriorsTab(QWidget* parent, Database* database)
-    : QWidget(parent), database_(database) {
+PosePriorsTab::PosePriorsTab(QWidget* parent) : QWidget(parent) {
   QGridLayout* grid = new QGridLayout(this);
 
   info_label_ = new QLabel(this);
@@ -755,7 +751,9 @@ PosePriorsTab::PosePriorsTab(QWidget* parent, Database* database)
   grid->setColumnStretch(0, 3);
 }
 
-void PosePriorsTab::Reload() {
+void PosePriorsTab::Reload(const std::shared_ptr<Database>& database) {
+  database_ = database;
+
   QString info;
   info += QString("Images: ") + QString::number(database_->NumImages());
   info += QString('\n');
@@ -838,9 +836,9 @@ DatabaseManagementWidget::DatabaseManagementWidget(QWidget* parent,
 
   tab_widget_ = new QTabWidget(this);
 
-  camera_tab_ = new CameraTab(this, &database_);
-  image_tab_ = new ImageTab(this, camera_tab_, options_, &database_);
-  pose_prior_tab_ = new PosePriorsTab(this, &database_);
+  camera_tab_ = new CameraTab(this);
+  image_tab_ = new ImageTab(this, camera_tab_, options_);
+  pose_prior_tab_ = new PosePriorsTab(this);
 
   tab_widget_->addTab(image_tab_, tr("Images"));
   tab_widget_->addTab(camera_tab_, tr("Cameras"));
@@ -870,11 +868,11 @@ DatabaseManagementWidget::DatabaseManagementWidget(QWidget* parent,
 void DatabaseManagementWidget::showEvent(QShowEvent*) {
   parent_->setDisabled(true);
 
-  database_.Open(*options_->database_path);
+  database_ = Database::Open(*options_->database_path);
 
-  image_tab_->Reload();
-  camera_tab_->Reload();
-  pose_prior_tab_->Reload();
+  image_tab_->Reload(database_);
+  camera_tab_->Reload(database_);
+  pose_prior_tab_->Reload(database_);
 }
 
 void DatabaseManagementWidget::hideEvent(QHideEvent*) {
@@ -884,7 +882,7 @@ void DatabaseManagementWidget::hideEvent(QHideEvent*) {
   camera_tab_->Clear();
   pose_prior_tab_->Clear();
 
-  database_.Close();
+  database_->Close();
 }
 
 void DatabaseManagementWidget::ClearMatches() {
@@ -893,10 +891,9 @@ void DatabaseManagementWidget::ClearMatches() {
                             "",
                             tr("Do you really want to clear all matches?"),
                             QMessageBox::Yes | QMessageBox::No);
-  if (reply == QMessageBox::No) {
-    return;
+  if (reply == QMessageBox::Yes) {
+    database_->ClearMatches();
   }
-  database_.ClearMatches();
 }
 
 void DatabaseManagementWidget::ClearTwoViewGeometries() {
@@ -905,10 +902,9 @@ void DatabaseManagementWidget::ClearTwoViewGeometries() {
       "",
       tr("Do you really want to clear all two-view geometries?"),
       QMessageBox::Yes | QMessageBox::No);
-  if (reply == QMessageBox::No) {
-    return;
+  if (reply == QMessageBox::Yes) {
+    database_->ClearTwoViewGeometries();
   }
-  database_.ClearTwoViewGeometries();
 }
 
 }  // namespace colmap
